+++ src/Commands/AbstractCommand.php
@@ -2,13 +2,16 @@
 
 namespace Rougin\Blueprint\Commands;
 
+use League\Flysystem\Filesystem;
+use Symfony\Component\Console\Command\Command;
+
 /**
  * Abstract Command
  *
  * @package Blueprint
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-abstract class AbstractCommand extends \Symfony\Component\Console\Command\Command
+abstract class AbstractCommand extends Command
 {
     /**
      * @var \League\Flysystem\Filesystem
@@ -21,14 +24,17 @@ abstract class AbstractCommand extends \Symfony\Component\Console\Command\Comman
     protected $renderer;
 
     /**
+     * Initializes the command instance.
+     *
      * @param \League\Flysystem\Filesystem $filesystem
      * @param \Twig_Environment            $renderer
      */
-    public function __construct(\League\Flysystem\Filesystem $filesystem, \Twig_Environment $renderer)
+    public function __construct(Filesystem $filesystem, \Twig_Environment $renderer)
     {
         parent::__construct();
 
         $this->filesystem = $filesystem;
-        $this->renderer   = $renderer;
+
+        $this->renderer = $renderer;
     }
 }